Update LICENSE
I (ROHIT SAXENA) created this code for EDUCATION PURPOSE ONLY. Any illegal or intentional harm or misuse of this code is strictly prohibited by the owner (Rohit Saxena GitHub ID: rohit290554) of this code. For any illegal or intentional harm or misuse of this code, the Owner of this code (Rohit Saxena GitHub ID: rohit290554) will not be responsible.